Target temperature may not be set in all ac modes

+++ bundles/org.openhab.binding.sensibo/src/test/java/org/openhab/binding/sensibo/internal/dto/GetPodDetailsResponseTest.java
@@ -96,7 +96,7 @@ public class GetPodDetailsResponseTest extends AbstractSerializationDeserializat
         assertTrue(acState.on);
         assertEquals("medium_high", acState.fanLevel);
         assertEquals("C", acState.temperatureUnit);
-        assertEquals(21, acState.targetTemperature);
+        assertEquals(21, acState.targetTemperature.intValue());
         assertEquals("heat", acState.mode);
         assertEquals("rangeFull", acState.swing);
